Moved BookDaoImpl class into the dao module from service, so it fits the pre-designed project structure.
