--- dao/src/main/java/hu/iit/uni/miskolc/swtest/App.java
@@ -1,13 +1,0 @@
-package hu.iit.uni.miskolc.swtest;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
